requirements.txt
streamlit
joblib
scikit-learn
pillow
numpy
